任务脚本

+++ src/Application.Shared/Items/ItemQuantity.cs
@@ -2,4 +2,4 @@ namespace Application.Shared.Items;
 
 public record ItemQuantity(int ItemId, int Quantity);
 
-public record TypedItemQuantity(int Type, ItemQuantity Item);
+public record TypedItemQuantity(sbyte Type, ItemQuantity Item);
